requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
bcrypt==4.1.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.2
discord==2.3.2
discord.py==2.3.2
frozenlist==1.4.1
idna==3.6
multidict==6.0.5
paramiko==3.4.0
pycparser==2.21
PyMySQL==1.1.0
PyNaCl==1.5.0
pysftp==0.2.9
python-dotenv==1.0.1
requests==2.31.0
urllib3==2.2.0
yarl==1.9.4
